feat: /kosen-setで学年を設定できるようにした
